replace MessengerModule with MessengerConfigurationRegistration
